add python-messages extensions
add control over generation of equality and comparison operators. in order for
one to be generated, add 'eq' or 'cmp' before the field in the structure
definition. Only the fields marked will be checked for the operators.

add control over generated serialisation functions, to allow a custom class to
subclass a generated class without needed to define reduntant serialise
functions. To do this, declare the custom class with:
'baseclass' subclass 'customclass' : <customclass.h>
currently only the existance of 'baseclass' is checked, not the value.

BoundedFloat and messages.msg were modified to demonstrate the extensions and so
that the build would finish sucessfully. They could probably be improves so that
the comparison operators have better definitions than they did before