switch to crossroads for zeromq messaging.
Still need zmq compatilibity for python stuff though

Crossroads is supposedly a more 'conservative' fork of zmq, but since the two of the
main zmq developers switched it's actually further along and less buggy than
current zmq git, so it's a better fit currently. It also has the benefit of a
released version which actually works.

This change is basically a sed -i 's/zmq/xs', but works better than trying to
used crossroad's C zmq compatibility layer because it's based on ZMQ 2 which has
a slightly different API.

The python stuff still uses pyzmq which should be built based on crossroad's
compatibility layer. The python API is the same so it's not a problem, just
confusing in name