Move serialisation to sending thread instead of monitoring thread
This is not as neat because it adds another lock, makes sendMessage potentially
block for a while if serialisation is expensive, and some other issues.

However, it means the message object is no longer destructed on the monitoring
thread, which boost::python *really* doesn't like (segfaults galore). This could
possibly be improved, but it's probably OK as it is