Monster commit. reorganise a bunch of headers into utility instead of common
There's now a 'utility' library which contains most of the cauv_util stuff and
other things with minimal dependencies. TimeStamp and UID now no longer depend
on message generation for their basic definitions, but are still serialisable
like before. Still working on getting the OpenCV dependency out of common (the
class was moved to special_messages)

This may be fun to merge into gui-graphicscene and other stuff