Refactor Image into BaseImage and Image to manage depencencies better
Image is now a subclass of BaseImage, which all that the messaging system knows
about. The virtual function bytes() is used to encode the image to JPEG before
being serialised. There's currently a lot more copies of the JPEG data than
necessary, and the JPEG encoding happens before the messaging system can decide
if it's needed or not, which will be fixed in a later commit

Linking changes will also come in the next commit