Get rid of some unnecessary threads. Synchronous now default for C++ node run
Main reason for change to synchronous as default is asynchronous appears to work
fine when used incorrectly (where synchronous would be more appropriate), but
produces a useless thread. If synchronous is used incorrectly then it's pretty
obvious (no messages). Asynchronous is still default for Python because it fits
in better with how CauvNode is used in practise there.

Ideally onRun could be removed entirely, but that's a bit outside the scope of
this patch