Remove unused mutex from cauv_debug which was causing gui abort on exit
Annoying details: the static initialisation of m_mutex is places in
cauv_debug.a, which is linked into both libwater.so and libgui-core.so. When
both these libraries are used in an executable, it gets constructed twice in the
same memory location, then destructed twice, causing a double free. The fix is
to hide the static initialisation in a function, but in this case it's easier to
just remove it.

see http://gcc.gnu.org/ml/gcc-help/2010-10/msg00248.html and replies